Add script to deploy Packer AMI with TF.
I was following a mix of two tutorials (both linked in README) which
caused me to mess up the SSH key for the Packer AMI I used.

This meant I couldn't start up the Go server mentioned in
https://learn.hashicorp.com/tutorials/terraform/packer?in=terraform/provision.

Something to follow on with in the future. Also need to build a more
sophisticated AMI next.